--- src/app/api/auth/[...nextauth]/route.js
@@ -1,42 +1,0 @@
-import NextAuth from "next-auth/next";
-import googleProvider from "next-auth/providers/google";
-import { FirestoreAdapter } from "@auth/firebase-adapter";
-import { cert } from "firebase-admin/app";
-
-export const authOptions = {
-  // eslint-disable-next-line new-cap
-  adapter: FirestoreAdapter({
-    credential: cert({
-      projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
-      clientEmail: process.env.NEXT_PUBLIC_FIREBASE_CLIENT_EMAIL,
-      privateKey: process.env.NEXT_PUBLIC_FIREBASE_PRIVATE_KEY,
-    }),
-  }),
-  providers: [
-    googleProvider({
-      profile(profile) {
-        return {
-          id: profile.sub,
-          email: profile.email,
-          name: profile.name,
-          image: profile.picture,
-          roles: {},
-        };
-      },
-      clientId: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID,
-      clientSecret: process.env.NEXT_PUBLIC_GOOGLE_CLIENT_SECRET,
-    }),
-  ],
-  pages: {
-    error: "/auth/error",
-  },
-  callbacks: {
-    async session({ session, user }) {
-      session.user = user;
-      return session;
-    },
-  },
-};
-// eslint-disable-next-line new-cap
-const handler = NextAuth(authOptions);
-export { handler as GET, handler as POST };
