+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,6 @@
+import NextAuth from "next-auth";
+import { options } from "@/utils/auth";
+
+// eslint-disable-next-line new-cap
+const handler = NextAuth(options);
+export { handler as GET, handler as POST };
