+++ src/app/layout.tsx
@@ -1,8 +1,9 @@
 /* eslint-disable new-cap */
 import "./globals.css";
 import { Poppins } from "next/font/google";
-import Session from "@/components/Session";
+import SessionProvider from "@/components/Session";
 import { Toaster } from "react-hot-toast";
+import { getServerSession } from "next-auth";
 
 const poppins = Poppins({
   subsets: ["latin"],
@@ -11,15 +12,21 @@ const poppins = Poppins({
   variable: "--font-poppins",
 });
 
-export const RootLayout = ({ children, session }) => {
+type Props = {
+  children: React.ReactNode;
+};
+
+const RootLayout = async ({ children }: Props) => {
+  const session = getServerSession();
+
   return (
     <html lang="en" className="h-full">
       <body className={`${poppins.variable} flex flex-col lg:flex-row h-full`}>
         <div className="flex w-full h-full">
-          <Session session={session} refetchInterval={5 * 60}>
+          <SessionProvider session={session}>
             <Toaster />
             {children}
-          </Session>
+          </SessionProvider>
         </div>
       </body>
     </html>
